board cards

+++ FrontEnd/src/app/Components/card-collection/card-collection.component.ts
@@ -2,7 +2,9 @@ import { Component, OnInit, Input } from '@angular/core';
 import { CardCollectionsService, CardCollection } from '../../Services/card-collections.service';
 import { CardsService, Card } from '../../Services/cards.service';
 import { HttpErrorResponse } from "@angular/common/http";
-import { MatSnackBar } from '@angular/material';
+import { ConfirmModalComponent } from '../../Modals/confirm-modal/confirm-modal.component';
+
+import { MatSnackBar, MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-card-collection',
@@ -16,29 +18,37 @@ export class CardCollectionComponent implements OnInit {
   cardDesc: string;
 
   constructor(private cardsService: CardsService, private cardCollectionsService: CardCollectionsService,
-    public snackBar: MatSnackBar) { }
+    public snackBar: MatSnackBar, public dialog: MatDialog) { }
 
   ngOnInit() {
   }
 
-  deleteCol(i: number){
-    this.cardCollectionsService.remove(this.collectionData[i]).subscribe((data: any) => {
-      console.log(data);
-      this.collectionData.splice(i, 1);
-    }, (err: HttpErrorResponse) => {
-      var errMsg = err.statusText + ': ';
-      Object.keys(err.error).forEach(function(e){
-        // errMsg += ' ' + e + ": "
-        err.error[e].forEach(function(str){
-          errMsg += str;
-        })
-      })
-      console.log(err);
-      this.openSnackBar(errMsg);
+  deleteCol(i: number) {
+    const dialogRef = this.dialog.open(ConfirmModalComponent, {
+      width: '500px'
     });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.cardCollectionsService.remove(this.collectionData[i]).subscribe((data: any) => {
+          console.log(data);
+          this.collectionData.splice(i, 1);
+        }, (err: HttpErrorResponse) => {
+          var errMsg = err.statusText + ': ';
+          Object.keys(err.error).forEach(function (e) {
+            // errMsg += ' ' + e + ": "
+            err.error[e].forEach(function (str) {
+              errMsg += str;
+            })
+          })
+          console.log(err);
+          this.openSnackBar(errMsg);
+        });
+      }
+    })
   }
 
-  deleteCard(collectionIndex: number, cardIndex: number){
+  deleteCard(collectionIndex: number, cardIndex: number) {
     var cardToBeDeleted = this.collectionData[collectionIndex].cards[cardIndex];
     console.log(cardToBeDeleted);
     this.collectionData[collectionIndex].cards.splice(cardIndex, 1);
@@ -47,7 +57,7 @@ export class CardCollectionComponent implements OnInit {
     })
   }
 
-  addCard(i:number, id: number){
+  addCard(i: number, id: number) {
     var c = new Card(this.cardTitle, this.cardDesc, id);
     this.cardDesc = '';
     this.cardTitle = '';
@@ -57,9 +67,9 @@ export class CardCollectionComponent implements OnInit {
       this.collectionData[i].cards.push(data);
     }, (err: HttpErrorResponse) => {
       var errMsg = err.statusText + ': ';
-      Object.keys(err.error).forEach(function(e){
+      Object.keys(err.error).forEach(function (e) {
         // errMsg += ' ' + e + ": "
-        err.error[e].forEach(function(str){
+        err.error[e].forEach(function (str) {
           errMsg += str;
         })
       })
